refactor(Scene Manager): Refactor the scene manager interface
The scene manager interface wasn't used and wasn't really started so I finished it. I also added some error classes
